+++ webdown/__init__.py
@@ -13,4 +13,4 @@ Basic usage:
 See the README.md file for more detailed usage instructions.
 """
 
-__version__ = "0.3.1"
+__version__ = "0.4.0"